1、修复在序列化时SkipSerialize注解未被解析的bug 2、升级erupt版本为版本升级做准备
