+++ erupt-annotation/pom.xml
@@ -13,7 +13,7 @@
     <parent>
         <groupId>xyz.erupt</groupId>
         <artifactId>erupt</artifactId>
-        <version>1.5.6</version>
+        <version>1.5.7</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
