+++ erupt-data/erupt-jpa/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>xyz.erupt</groupId>
         <artifactId>erupt</artifactId>
-        <version>1.5.6</version>
+        <version>1.5.7</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
