+++ erupt-data/erupt-mongodb/pom.xml
@@ -10,7 +10,7 @@
     <parent>
         <groupId>xyz.erupt</groupId>
         <artifactId>erupt</artifactId>
-        <version>1.5.6</version>
+        <version>1.5.7</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
