+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>xyz.erupt</groupId>
     <artifactId>erupt</artifactId>
-    <version>1.5.6</version>
+    <version>1.5.7</version>
 
     <packaging>pom</packaging>
     <name>erupt</name>
